backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
cryptography==46.0.5
distro==1.9.0
dj-database-url==3.1.0
Django==4.2.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==23.5
django-ratelimit==4.1.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.26.5
google-auth==2.48.0
google-genai==1.63.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
numpy==2.4.2
packaging==26.0
pgvector==0.4.2
pillow==12.1.1
psycopg==3.3.2
psycopg-binary==3.3.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
setuptools==80.10.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
wheel==0.46.3
whitenoise==6.6.0
